requirements.txt
tensorflow-cpu
tflearn~=0.5.0
pympler
cvxopt~=1.2.6
seaborn
pandas~=1.3.1
yfinance~=0.1.63
Pillow
setuptools~=57.0.0
numpy~=1.19.5
scipy~=1.7.0
matplotlib~=3.4.2
yahoo-fin
beautifulsoup4~=4.9.3